refactor: move to MonthFootersDemoViewController
